Set TZ to GMT when using -i or -s flags to standardise calculations
Modify the `parse_options` function to set the TZ environment variable
to "GMT" exclusively when either of the -i (incremental) or -s (since
start) flags are used. This change ensures that the timestamps generated
are consistent and not influenced by local time zone shifts, mirroring
the behaviour of the original `ts` program from `moreutils`.

I also removed the previous checking and setting of TZ in the `main`
function, which was unnecessary.